Craft Update #1

+++ craft/app/services/AssetTransformsService.php
@@ -2,27 +2,38 @@
 namespace Craft;
 
 /**
- * Craft by Pixel & Tonic
+ * Class AssetTransformsService
  *
- * @package   Craft
- * @author    Pixel & Tonic, Inc.
+ * @author    Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @copyright Copyright (c) 2014, Pixel & Tonic, Inc.
  * @license   http://buildwithcraft.com/license Craft License Agreement
- * @link      http://buildwithcraft.com
- */
-
-/**
- *
+ * @see       http://buildwithcraft.com
+ * @package   craft.app.services
+ * @since     1.0
  */
 class AssetTransformsService extends BaseApplicationComponent
 {
+	// Properties
+	// =========================================================================
+
+	/**
+	 * @var
+	 */
 	private $_transformsByHandle;
+
+	/**
+	 * @var bool
+	 */
 	private $_fetchedAllTransforms = false;
+
+	// Public Methods
+	// =========================================================================
 
 	/**
 	 * Returns all named asset transforms.
 	 *
 	 * @param string|null $indexBy
+	 *
 	 * @return array
 	 */
 	public function getAllTransforms($indexBy = null)
@@ -66,13 +77,14 @@ class AssetTransformsService extends BaseApplicationComponent
 	/**
 	 * Returns an asset transform by its handle.
 	 *
-	 * @param $handle
+	 * @param string $handle
+	 *
 	 * @return AssetTransformModel|null
 	 */
 	public function getTransformByHandle($handle)
 	{
-		// If we've already fetched all transforms we can save ourselves a trip to the DB
-		// for transform handles that don't exist
+		// If we've already fetched all transforms we can save ourselves a trip to the DB for transform handles that
+		// don't exist
 		if (!$this->_fetchedAllTransforms &&
 			(!isset($this->_transformsByHandle) || !array_key_exists($handle, $this->_transformsByHandle))
 		)
@@ -99,13 +111,13 @@ class AssetTransformsService extends BaseApplicationComponent
 		}
 	}
 
-
 	/**
 	 * Saves an asset transform.
 	 *
 	 * @param AssetTransformModel $transform
-	 * @return bool
+	 *
 	 * @throws Exception
+	 * @return bool
 	 */
 	public function saveTransform(AssetTransformModel $transform)
 	{
@@ -166,6 +178,7 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Deletes an asset transform by it's id.
 	 *
 	 * @param int $transformId
+	 *
 	 * @throws \Exception
 	 * @return bool
 	 */
@@ -178,8 +191,10 @@ class AssetTransformsService extends BaseApplicationComponent
 	/**
 	 * Update the asset transforms for the FileModel.
 	 *
-	 * @param AssetFileModel $fileModel
+	 * @param AssetFileModel      $fileModel
 	 * @param array|object|string $transformsToUpdate
+	 *
+	 * @deprecated This method will likely be modified or removed in a future update.
 	 * @return bool
 	 */
 	public function updateTransforms(AssetFileModel $fileModel, $transformsToUpdate)
@@ -242,7 +257,10 @@ class AssetTransformsService extends BaseApplicationComponent
 			}
 		}
 
-		$this->deleteSourceIfNecessary($imageSource);
+		if (craft()->assetSources->populateSourceType($fileModel->getSource())->isRemote())
+		{
+			$this->deleteSourceIfNecessary($imageSource);
+		}
 
 		return true;
 	}
@@ -251,7 +269,8 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Get a transform index row. If it doesn't exist - create one.
 	 *
 	 * @param AssetFileModel $file
-	 * @param $transform
+	 * @param                $transform
+	 *
 	 * @return AssetTransformIndexModel
 	 */
 	public function getTransformIndex(AssetFileModel $file, $transform)
@@ -304,8 +323,9 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Get a transform URL by the transform index model.
 	 *
 	 * @param AssetTransformIndexModel $transformIndexModel
-	 * @return string
+	 *
 	 * @throws Exception
+	 * @return string
 	 */
 	public function ensureTransformUrlByIndexModel(AssetTransformIndexModel $transformIndexModel)
 	{
@@ -380,6 +400,8 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Index a transform.
 	 *
 	 * @param AssetTransformIndexModel $transformIndexData
+	 *
+	 * @deprecated This method will likely be modified or removed in a future update.
 	 * @return bool
 	 */
 	public function generateTransform(AssetTransformIndexModel $transformIndexData)
@@ -429,8 +451,8 @@ class AssetTransformsService extends BaseApplicationComponent
 			{
 				if (!$transform->isNamedTransform() || ($transform->isNamedTransform() && $existingFileTimeModified >= $transform->dimensionChangeTime))
 				{
-					// We have a satisfactory match and the record has been inserted already.
-					// Now copy the file to the new home
+					// We have a satisfactory match and the record has been inserted already. Now copy the file to the
+					// new home.
 					$sourceType->copyTransform($file, $alternateLocation, $transformIndexData->location);
 					return true;
 				}
@@ -446,6 +468,8 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Get a transform's subpath
 	 *
 	 * @param $transform
+	 *
+	 * @deprecated This method will likely be modified or removed in a future update.
 	 * @return string
 	 */
 	public function getTransformSubpath($transform)
@@ -457,8 +481,9 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Normalize a transform from handle or a set of properties to an AssetTransformModel.
 	 *
 	 * @param mixed $transform
-	 * @return AssetTransformModel|null
+	 *
 	 * @throws Exception
+	 * @return AssetTransformModel|null
 	 */
 	public function normalizeTransform($transform)
 	{
@@ -469,6 +494,7 @@ class AssetTransformsService extends BaseApplicationComponent
 		else if (is_string($transform))
 		{
 			$transformModel =  $this->getTransformByHandle($transform);
+
 			if ($transformModel)
 			{
 				return $transformModel;
@@ -493,6 +519,7 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Store a transform index data by it's model.
 	 *
 	 * @param AssetTransformIndexModel $data
+	 *
 	 * @return AssetTransformIndexModel
 	 */
 	public function storeTransformIndexData(AssetTransformIndexModel $data)
@@ -514,7 +541,8 @@ class AssetTransformsService extends BaseApplicationComponent
 	/**
 	 * Get a transform index model by a row id.
 	 *
-	 * @param $transformId
+	 * @param int $transformId
+	 *
 	 * @return AssetTransformIndexModel|null
 	 */
 	public function getTransformIndexModelById($transformId)
@@ -537,8 +565,9 @@ class AssetTransformsService extends BaseApplicationComponent
 	/**
 	 * Get a transform index model by a row id.
 	 *
-	 * @param $fileId
-	 * @param $transformHandle
+	 * @param int    $fileId
+	 * @param string $transformHandle
+	 *
 	 * @return AssetTransformIndexModel|null
 	 */
 	public function getTransformIndexModelByFileIdAndHandle($fileId, $transformHandle)
@@ -547,7 +576,7 @@ class AssetTransformsService extends BaseApplicationComponent
 		$entry =  craft()->db->createCommand()
 			->select('ti.*')
 			->from('assettransformindex ti')
-			->where('ti.fileId = :id AND ti.location = :location', array(':id' => $fileId, ':location' => '_' . $transformHandle))
+			->where('ti.fileId = :id AND ti.location = :location', array(':id' => $fileId, ':location' => '_'.$transformHandle))
 			->queryRow();
 
 		if ($entry)
@@ -561,7 +590,8 @@ class AssetTransformsService extends BaseApplicationComponent
 	/**
 	 * Get URL for Transform by TransformIndexId.
 	 *
-	 * @param $transformId
+	 * @param int $transformId
+	 *
 	 * @return string
 	 */
 	public function getUrlforTransformByIndexId($transformId)
@@ -579,7 +609,9 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Get URL for a transform by File Model and transform.
 	 *
 	 * @param AssetFileModel $file
-	 * @param $transform
+	 * @param                $transform
+	 *
+	 * @deprecated This method will likely be modified or removed in a future update.
 	 * @return string
 	 */
 	public function getUrlforTransformByFile($file, $transform)
@@ -587,31 +619,35 @@ class AssetTransformsService extends BaseApplicationComponent
 		// Create URL to the image
 		$sourceType = craft()->assetSources->getSourceTypeById($file->sourceId);
 		$transformPath = $this->getTransformSubpath($transform);
+
 		return AssetsHelper::generateUrl($sourceType, $file, $transformPath);
 	}
 
 	/**
 	 * Cleans up transforms for a source by making sure that all indexed transforms actually exist.
 	 *
-	 * @param $sourceId
+	 * @param int $sourceId
+	 *
+	 * @return null
 	 */
 	public function cleanUpTransformsForSource($sourceId)
 	{
 		$transformList = craft()->db->createCommand()
-							->where('sourceId = :sourceId AND fileExists = 1', array(':sourceId' => $sourceId))
-							->select('*')
-							->from('assettransformindex')
-							->queryAll();
+		        ->where('sourceId = :sourceId AND fileExists = 1', array(':sourceId' => $sourceId))
+		        ->select('*')
+		        ->from('assettransformindex')
+		        ->queryAll();
 
 		$sourceType = craft()->assetSources->getSourceTypeById($sourceId);
 
 		foreach ($transformList as $row)
 		{
-				$file = craft()->assets->getFileById($row['fileId']);
-				if (!$file || !$sourceType->transformExists($file, $row['location']))
-				{
-					craft()->db->createCommand()->delete('assettransformindex', 'id = '.$row['id']);
-				}
+			$file = craft()->assets->getFileById($row['fileId']);
+
+			if (!$file || !$sourceType->transformExists($file, $row['location']))
+			{
+				craft()->db->createCommand()->delete('assettransformindex', 'id = '.$row['id']);
+			}
 		}
 	}
 
@@ -619,6 +655,7 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Get generated transform locations for a file.
 	 *
 	 * @param AssetFileModel $file
+	 *
 	 * @return array|\CDbDataReader
 	 */
 	public function getGeneratedTransformLocationsForFile(AssetFileModel $file)
@@ -634,7 +671,9 @@ class AssetTransformsService extends BaseApplicationComponent
 	/**
 	 * Delete transform records by a file id.
 	 *
-	 * @param $fileId
+	 * @param int $fileId
+	 *
+	 * @return null
 	 */
 	public function deleteTransformRecordsByFileId($fileId)
 	{
@@ -646,9 +685,10 @@ class AssetTransformsService extends BaseApplicationComponent
 	 *
 	 * @param $fileModel
 	 * @param $size
+	 *
 	 * @return bool|string
 	 */
-	public function getThumbServerPath($fileModel, $size)
+	public function getThumbServerPath(AssetFileModel $fileModel, $size)
 	{
 		$thumbFolder = craft()->path->getAssetsThumbsPath().$size.'/';
 		IOHelper::ensureFolderExists($thumbFolder);
@@ -663,76 +703,22 @@ class AssetTransformsService extends BaseApplicationComponent
 				->scaleAndCrop($size, $size)
 				->saveAs($thumbPath);
 
-			$this->deleteSourceIfNecessary($imageSource);
+			if (craft()->assetSources->populateSourceType($fileModel->getSource())->isRemote())
+			{
+				$this->deleteSourceIfNecessary($imageSource);
+			}
 		}
 
 		return $thumbPath;
-	}
-
-	// Private methods
-
-	/**
-	 * Returns a DbCommand object prepped for retrieving transforms.
-	 *
-	 * @return DbCommand
-	 */
-	private function _createTransformQuery()
-	{
-		return craft()->db->createCommand()
-			->select('id, name, handle, mode, position, height, width, quality, dimensionChangeTime')
-			->from('assettransforms')
-			->order('name');
-	}
-
-	/**
-	 * Returns a trasnform's folder name.
-	 *
-	 * @param AssetTransformModel $transform
-	 * @return string
-	 */
-	private function _getTransformFolderName(AssetTransformModel $transform)
-	{
-		if ($transform->isNamedTransform())
-		{
-			return $this->_getNamedTransformFolderName($transform);
-		}
-		else
-		{
-			return $this->_getUnnamedTransformFolderName($transform);
-		}
-	}
-
-	/**
-	 * Returns a named transform's folder name.
-	 *
-	 * @param AssetTransformModel $transform
-	 * @return string
-	 */
-	private function _getNamedTransformFolderName(AssetTransformModel $transform)
-	{
-		return '_'.$transform->handle;
-	}
-
-	/**
-	 * Returns an unnamed transform's folder name.
-	 *
-	 * @param AssetTransformModel $transform
-	 * @return string
-	 */
-	private function _getUnnamedTransformFolderName(AssetTransformModel $transform)
-	{
-		return '_'.($transform->width ? $transform->width : 'AUTO').'x'.($transform->height ? $transform->height : 'AUTO') .
-		       '_'.($transform->mode) .
-		       '_'.($transform->position) .
-		       ($transform->quality ? '_'.$transform->quality : '');
 	}
 
 	/**
 	 * Get a local image source to use for transforms.
 	 *
 	 * @param $fileModel
-	 * @return mixed
+	 *
 	 * @throws Exception
+	 * @return mixed
 	 */
 	public function getLocalImageSource($fileModel)
 	{
@@ -767,6 +753,9 @@ class AssetTransformsService extends BaseApplicationComponent
 	 * Deletes an image local source if required by config.
 	 *
 	 * @param $imageSource
+	 *
+	 * @deprecated This method will likely be modified or removed in a future update.
+	 * @return null
 	 */
 	public function deleteSourceIfNecessary($imageSource)
 	{
@@ -781,6 +770,8 @@ class AssetTransformsService extends BaseApplicationComponent
 	 *
 	 * @param $localCopy
 	 * @param $destination
+	 *
+	 * @return null
 	 */
 	public function storeLocalSource($localCopy, $destination)
 	{
@@ -790,11 +781,80 @@ class AssetTransformsService extends BaseApplicationComponent
 		if ($maxCachedImageSize > 0)
 		{
 			craft()->images->loadImage($localCopy)->scaleToFit($maxCachedImageSize, $maxCachedImageSize)->setQuality(100)->saveAs($destination);
-			IOHelper::deleteFile($localCopy);
+
+			if ($localCopy != $destination)
+			{
+				IOHelper::deleteFile($localCopy);
+			}
 		}
 		else
 		{
-			IOHelper::move($localCopy, $destination);
+			if ($localCopy != $destination)
+			{
+				IOHelper::move($localCopy, $destination);
+			}
 		}
+	}
+
+	// Private Methods
+	// =========================================================================
+
+	/**
+	 * Returns a DbCommand object prepped for retrieving transforms.
+	 *
+	 * @return DbCommand
+	 */
+	private function _createTransformQuery()
+	{
+		return craft()->db->createCommand()
+			->select('id, name, handle, mode, position, height, width, quality, dimensionChangeTime')
+			->from('assettransforms')
+			->order('name');
+	}
+
+	/**
+	 * Returns a transform's folder name.
+	 *
+	 * @param AssetTransformModel $transform
+	 *
+	 * @return string
+	 */
+	private function _getTransformFolderName(AssetTransformModel $transform)
+	{
+		if ($transform->isNamedTransform())
+		{
+			return $this->_getNamedTransformFolderName($transform);
+		}
+		else
+		{
+			return $this->_getUnnamedTransformFolderName($transform);
+		}
+	}
+
+	/**
+	 * Returns a named transform's folder name.
+	 *
+	 * @param AssetTransformModel $transform
+	 *
+	 * @return string
+	 */
+	private function _getNamedTransformFolderName(AssetTransformModel $transform)
+	{
+		return '_'.$transform->handle;
+	}
+
+	/**
+	 * Returns an unnamed transform's folder name.
+	 *
+	 * @param AssetTransformModel $transform
+	 *
+	 * @return string
+	 */
+	private function _getUnnamedTransformFolderName(AssetTransformModel $transform)
+	{
+		return '_'.($transform->width ? $transform->width : 'AUTO').'x'.($transform->height ? $transform->height : 'AUTO') .
+		       '_'.($transform->mode) .
+		       '_'.($transform->position) .
+		       ($transform->quality ? '_'.$transform->quality : '');
 	}
 }
